Set facter implementation to rspec
This completely disables Facter and uses a stubbed implementation. This
is fine since we use use rspec-puppet-facts to set all facts. It gives
complete isolation from the host system and is also significantly faster
with Facter 4 (which always also resolves the facts, see [FACT-3039]).
It only works with Puppet 6.25+ and 7.12+ and requires rspec-puppet
2.11.0+.

FACT-3039: https://tickets.puppetlabs.com/browse/FACT-3039