Correct systemd custom fact implementation
Since rspec-puppet-facts 1.9.5 (and in particular
21b679f45b9e20902af53161b5e1bd0b4fb723f7[1]) custom facts are stored as
strings. Since this gem requires rspec-puppet-facts >= 2.0.1 they should
be treated as such.

This also highlighted that we're incompatible with
RSpec.configuration.facterdb_string_keys set to true, but most testing
code in Vox Pupuli is incompatible with this.

Tests are added to ensure it really works and prevents future
regressions.

[1]: https://github.com/mcanevet/rspec-puppet-facts/commit/21b679f45b9e20902af53161b5e1bd0b4fb723f7